requirements.txt
asgiref==3.2.10
Django==3.0
django-cors-headers==3.1.1
django-polymorphic==3.1.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1